tests/wpt/tests/html/canvas/offscreen/compositing/2d.composite.globalAlpha.image.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.composite.globalAlpha.image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.composite.globalAlpha.image</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  // Avoiding any potential alpha = 0 optimisations.
  ctx.globalAlpha = 0.01;
  const response = await fetch('/images/red.png');
  const blob = await response.blob();
  const bitmap = await createImageBitmap(blob);

  ctx.drawImage(bitmap, 0, 0);
  _assertPixelApprox(canvas, 50,25, 2,253,0,255, 2);

}, "");
</script>
